# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EncoreHub Gateway API",
    "version": "0.1.0",
    "description": "Browser-facing HTTP and SSE contract used by the EncoreHub frontend. Engine-only routes and third-party provider APIs are outside this contract. Every response includes X-Request-ID."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Default headless development address; desktop ports are negotiated at runtime."
    }
  ],
  "security": [
    {},
    {
      "GatewayBearer": []
    }
  ],
...
  "paths": {
...
    "/api/v1/conversations": {
      "get": {
        "tags": ["Conversations"],
        "operationId": "listConversations",
        "responses": {
          "200": { "$ref": "#/components/responses/ConversationList" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Conversations"],
        "operationId": "createConversation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateConversationRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Conversation" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" }
      ],
      "get": {
        "tags": ["Conversations"],
        "operationId": "getConversation",
        "responses": {
          "200": { "$ref": "#/components/responses/ConversationDetail" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Conversations"],
        "operationId": "updateConversation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateConversationRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Conversation" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Conversations"],
        "operationId": "deleteConversation",
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/chat": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" },
        { "$ref": "#/components/parameters/ProviderKey" }
      ],
      "post": {
        "tags": ["Conversations"],
        "operationId": "sendChatMessage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response when stream=false, otherwise an SSE stream ending in authoritative done or error data. Streaming delta, reasoning, and usage events may include cumulative provider generation duration_ms; tool execution time is excluded.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatResponse" }
              },
              "text/event-stream": {
                "schema": { "type": "string" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/attachments": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" }
      ],
      "get": {
        "tags": ["Attachments"],
        "operationId": "listConversationAttachments",
        "responses": {
          "200": {
            "description": "Attachment metadata in upload order",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Attachments"],
        "operationId": "uploadConversationAttachment",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": { "file": { "type": "string", "format": "binary" } }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Attachment" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/conversations/{id}/generate-title": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" },
        { "$ref": "#/components/parameters/ProviderKey" }
      ],
      "post": {
        "tags": ["Conversations"],
        "operationId": "generateConversationTitle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GenerateTitleRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Conversation" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/title": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" }
      ],
      "patch": {
        "tags": ["Conversations"],
        "operationId": "updateConversationTitle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RenameConversationRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Conversation" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/messages/{message_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" },
        { "name": "message_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "delete": {
        "tags": ["Conversations"],
        "operationId": "deleteConversationMessage",
        "responses": {
          "204": { "$ref": "#/components/responses/NoContent" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/conversations/{id}/character-upgrade": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" }
      ],
      "get": {
        "tags": ["Characters"],
        "operationId": "previewConversationCharacterUpgrade",
        "responses": {
          "200": { "$ref": "#/components/responses/CharacterUpgradePreview" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Characters"],
        "operationId": "upgradeConversationCharacter",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpgradeCharacterRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Conversation" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/search": {
      "post": {
        "tags": ["Search"],
        "operationId": "searchWeb",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SearchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SearchResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/log-level": {
      "get": {
        "tags": ["Settings"],
        "operationId": "getLogLevel",
        "responses": {
          "200": {
            "description": "Persisted runtime log level",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LogLevelResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Settings"],
        "operationId": "setLogLevel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LogLevelRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applied runtime log level",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LogLevelResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/providers": {
      "get": {
        "tags": ["Providers"],
        "operationId": "listProviders",
        "responses": {
          "200": { "$ref": "#/components/responses/Providers" }
        }
      },
      "put": {
        "tags": ["Providers"],
        "operationId": "replaceProviders",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["providers"],
                "properties": {
                  "providers": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/ProviderProfile" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Providers" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/providers/{provider}/validate-key": {
      "parameters": [
        { "$ref": "#/components/parameters/ProviderId" },
        { "$ref": "#/components/parameters/ProviderKey" }
      ],
      "post": {
        "tags": ["Providers"],
        "operationId": "validateProviderKey",
        "description": "Validate request-local legacy or versioned API-key pools against draft endpoints without persisting keys or provider profiles. Responses contain only opaque key/endpoint IDs, latency, and structured error categories.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProviderKeyValidationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Per-key credential status and per-endpoint connection health",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderKeyValidationResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/providers/{provider}/models": {
      "parameters": [
        { "$ref": "#/components/parameters/ProviderId" },
        { "$ref": "#/components/parameters/ProviderKey" }
      ],
      "get": {
        "tags": ["Providers"],
        "operationId": "listProviderModels",
        "responses": {
          "200": {
            "description": "Provider models",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelsResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/providers/{provider}/embeddings": {
      "parameters": [
        { "$ref": "#/components/parameters/ProviderId" },
        { "$ref": "#/components/parameters/ProviderKey" }
      ],
      "post": {
        "tags": ["Providers"],
        "operationId": "createProviderEmbeddings",
        "description": "Create OpenAI-compatible float embeddings as a standalone utility operation. Embedding calls do not create, modify, or participate in conversations.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEmbeddingsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OpenAI-compatible embedding vectors",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmbeddingsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/providers/{provider}/models/discover": {
      "parameters": [
        { "$ref": "#/components/parameters/ProviderId" },
        { "$ref": "#/components/parameters/ProviderKey" }
      ],
      "post": {
        "tags": ["Providers"],
        "operationId": "discoverProviderModels",
        "description": "Probe draft endpoints with a legacy single key or versioned API-key pool and return remote model candidates without persisting the profile or keys.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModelDiscoveryRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Structured discovery result, including per-endpoint failures",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelDiscoveryResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "GatewayBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required only when ENCOREHUB_AUTH_TOKEN is configured. Never reuse ENCOREHUB_ENGINE_AUTH_TOKEN."
      }
    },
    "parameters": {
      "ConversationId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
      "ProviderId": {
        "name": "provider",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "ProviderKey": {
        "name": "X-Provider-Key",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Optional when the provider secret is available from the unlocked Engine vault. Accepts a legacy single key or EncoreHub's versioned API-key pool envelope; values are never logged or persisted in ProviderProfile."
      }
    },
...
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NoContent": { "description": "Operation completed" },
...
      "Conversation": {
        "description": "Conversation summary",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Conversation" }
          }
        }
      },
      "ConversationDetail": {
        "description": "Conversation and authoritative messages",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ConversationDetail" }
          }
        }
      },
      "ConversationList": {
        "description": "Conversation summaries",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ConversationList" }
          }
        }
      },
      "Attachment": {
        "description": "Conversation-owned attachment metadata. storage_path is relative to the Engine data directory; original bytes are stored outside SQLite.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Attachment" }
          }
        }
      },
...
      "CharacterUpgradePreview": {
        "description": "Diff between the conversation snapshot and the character's current version",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/CharacterUpgradePreview" }
          }
        }
      },
      "Providers": {
        "description": "Canonical provider profiles",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ProvidersResponse" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
...
      "Attachment": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "conversation_id", "message_id", "file_name", "mime_type", "file_category", "size_bytes", "sha256", "storage_path", "processing_status", "processing_method", "extracted_text", "error_message", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "conversation_id": { "type": "string", "format": "uuid" },
          "message_id": { "type": ["string", "null"], "format": "uuid" },
          "file_name": { "type": "string" },
          "mime_type": { "type": "string" },
          "file_category": { "enum": ["image", "rich_text", "text"] },
          "size_bytes": { "type": "integer", "minimum": 0, "maximum": 20971520 },
          "sha256": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
          "storage_path": { "type": "string", "description": "Content-addressed path relative to the Engine data directory" },
          "processing_status": { "enum": ["pending", "ready", "failed"] },
          "processing_method": { "type": "string" },
          "extracted_text": { "type": "string" },
          "error_message": { "type": "string" },
          "created_at": { "type": "integer", "description": "Unix epoch milliseconds" },
          "updated_at": { "type": "integer", "description": "Unix epoch milliseconds" }
        }
      },
...
      "CreateConversationRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "provider": { "type": "string" },
          "model": { "type": "string" },
          "character_id": {
            "type": "string",
            "description": "Defaults to the stable default character. Provider/model inherit from the selected character when omitted together."
          }
        }
      },
...
      "UpgradeCharacterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expected_character_version"],
        "properties": {
          "expected_character_version": { "type": "integer", "minimum": 1 }
        }
      },
      "RenameConversationRequest": {
        "type": "object",
        "required": ["title"],
        "properties": { "title": { "type": "string", "minLength": 1 } }
      },
      "UpdateConversationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": { "type": "string", "minLength": 1 },
          "provider": { "type": "string", "minLength": 1 },
          "model": { "type": "string", "minLength": 1 }
        },
        "anyOf": [
          { "required": ["title"] },
          { "required": ["provider", "model"] }
        ]
      },
      "ChatRequest": {
        "type": "object",
        "properties": {
          "content": { "type": "string" },
          "provider": { "type": "string" },
          "model": { "type": "string" },
          "stream": { "type": "boolean" },
          "search": { "type": "boolean" },
          "search_provider": { "enum": ["duckduckgo", "bing", "google", "custom"] },
          "temperature": { "type": "number" },
          "top_p": { "type": "number" },
          "max_tokens": { "type": "integer" },
          "max_completion_tokens": { "type": "integer" },
          "frequency_penalty": { "type": "number" },
          "presence_penalty": { "type": "number" },
          "stop": { "type": "array", "items": { "type": "string" } },
          "seed": { "type": ["integer", "null"] },
          "json_mode": { "type": "boolean" },
          "reasoning_effort": { "type": "string", "enum": ["low", "medium", "high"] },
          "thinking_budget": { "type": "integer", "minimum": 0 },
          "attachment_ids": { "type": "array", "maxItems": 10, "items": { "type": "string", "format": "uuid" } },
          "model_supports_vision": { "type": "boolean", "description": "Whether the selected generation model accepts native image parts" },
          "image_strategy": { "enum": ["direct", "system_ocr", "vision_model"], "description": "Explicit user choice for image handling; models are never switched silently" },
          "vision_provider": { "type": "string" },
          "vision_model": { "type": "string" },
          "user_system_context": { "$ref": "#/components/schemas/UserSystemContext" },
          "replace_message_id": {
            "type": "string",
            "description": "User message ID to replace before regenerating the conversation from that turn"
          }
        },
        "anyOf": [
          { "required": ["content"], "properties": { "content": { "minLength": 1 } } },
          { "required": ["attachment_ids"], "properties": { "attachment_ids": { "minItems": 1 } } }
        ]
      },
...
      "ChatResponse": {
        "type": "object",
        "required": ["conversation_id", "reply", "provider", "model"],
        "properties": {
          "conversation_id": { "type": "string", "format": "uuid" },
          "user_message": { "$ref": "#/components/schemas/Message" },
          "assistant_message": { "$ref": "#/components/schemas/Message" },
          "reply": { "type": "string" },
          "provider": { "type": "string" },
          "model": { "type": "string" },
          "usage": { "$ref": "#/components/schemas/Usage" }
        }
      },
      "GenerateTitleRequest": {
        "type": "object",
        "properties": { "force": { "type": "boolean", "default": false } }
      },
      "SearchRequest": {
        "type": "object",
        "required": ["query"],
        "properties": {
          "query": { "type": "string", "minLength": 1, "maxLength": 500 },
          "max_results": { "type": "integer", "minimum": 1, "maximum": 10 },
          "provider": { "enum": ["duckduckgo", "bing", "google", "custom"] }
        }
      },
...
      "SearchResponse": {
        "type": "object",
        "required": ["query", "provider", "results"],
        "properties": {
          "query": { "type": "string" },
          "provider": { "type": "string" },
          "results": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SearchResult" }
          }
        }
      },
      "LogLevelRequest": {
        "type": "object",
        "required": ["level"],
        "properties": {
          "level": { "enum": ["error", "warn", "warning", "info", "debug", "trace"] }
        }
      },
      "LogLevelResponse": {
        "type": "object",
        "required": ["level"],
        "properties": {
          "level": { "type": "string" },
          "engine_applied": { "type": "boolean" }
        }
      },
      "ProviderProfile": {
        "type": "object",
        "required": ["id", "name", "protocol", "base_url", "models", "enabled", "builtin"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "protocol": { "enum": ["openai", "anthropic"] },
          "base_url": { "type": "string" },
          "models": { "type": "array", "items": { "type": "string" } },
          "endpoints": {
            "type": "array",
            "maxItems": 16,
            "items": { "$ref": "#/components/schemas/ProviderEndpoint" }
          },
          "routing_strategy": { "enum": ["failover", "round_robin"] },
          "key_routing_strategy": {
            "enum": ["failover", "round_robin"],
            "description": "Selection policy for the separately encrypted API-key pool."
          },
          "model_configs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProviderModelConfig" }
          },
          "enabled": { "type": "boolean" },
          "builtin": { "type": "boolean" }
        }
      },
...
      "ModelsResponse": {
        "type": "object",
        "required": ["provider", "models"],
        "properties": {
          "provider": { "type": "string" },
          "models": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ModelInfo" }
          }
        }
      },
      "ProviderKeyValidationRequest": {
        "type": "object",
        "required": ["protocol", "endpoints"],
        "properties": {
          "protocol": { "enum": ["openai", "anthropic"] },
          "endpoints": {
            "type": "array",
            "minItems": 1,
            "maxItems": 16,
            "items": { "$ref": "#/components/schemas/ProviderEndpoint" }
          }
        }
      },
...
    }
  }
}